Add license details to OpenAPI schema
This would make it easier for 3rd parties to integrate with Stalwart by being clear about licensing. I'm not sure if you want this license or want to release this file specifically under a more permissive license (MIT license would make sense here in my opinion, because it's just an API schema).