Allow to switch to operating system CA root store
So far, we only used the root CA certificates from the "webpki-roots"
crate, which makes it very difficult if you want to run a custom CA.

In my case I'm running automated tests of full production systems, which
also includes an ACME setup that injects a Pebble instance into the
system's CA certificates.

Since Stalwart doesn't use the system's root CA certificate store, it's
very difficult to inject the certificate for the Pebble instance.

Given that there's also some interest (issue #247) for doing this in an
enterprise environment with intranet CAs, I decided to generalise this
far enough that it can not only be used in my downstream setup.

Currently, this doesn't fully address the issue, since in the long term
this might be something we'd want to configure at runtime, as per
@mdecimus's comment[1]:

> A slightly more complex approach is required that allows the user to
> select at runtime which CA store to use.

However, this makes it at least easier to switch to native root CA store
by simply recompiling with the "tls-native-roots" feature.

[1]: https://github.com/stalwartlabs/stalwart/issues/247#issuecomment-2437039500

Signed-off-by: aszlig <aszlig@nix.build>
Issue: https://github.com/stalwartlabs/stalwart/issues/247